requirements.txt
numpy
pandas
mlflow==2.2.2
notebook
matplotlib
seaborn
scikit-learn
ython-box==7.3.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-cors
feature-engine
pycountry -q
category-encoders
feature-engine
python-dotenv
-e .